moved put part to aws request folder
