+++ src/FileUpload/FileUpload.ts
@@ -1,8 +1,8 @@
-import { PutPart } from './PutPart'
 import { CompleteMultipartUpload } from '../MultipartUpload/CompleteMultipartUpload'
 import { InitiateMultipartUpload } from '../MultipartUpload/InitiateMultipartUpload'
 import { DeleteMultipartUpload } from '../MultipartUpload/DeleteMultipartUpload'
 import { ResumeInterruptedUpload } from '../AwsRequest/ResumeInterruptedUpload'
+import { PutPart } from '../AwsRequest/PutPart'
 import { ReuseS3Object } from './ReuseS3Object'
 import { Global } from '../Global'
 import { FAR_FUTURE, ETAG_OF_0_LENGTH_BLOB } from '../Constants'
